rename css-carousel to carousel
