+++ src/index.ts
@@ -4,4 +4,4 @@ export * from './icon/Close'
 export * from './form/CheckBox'
 export * from './form/RadioInput'
 export * from './modal/modal'
-export * from './css-carousel'
+export * from './carousel'
