+++ stories/carousel.stories.js
@@ -1,30 +1,28 @@
 import React from 'react'
-
+import styled from 'styled-components'
 import { storiesOf } from '@storybook/react'
 import { action } from '@storybook/addon-actions'
 import { linkTo } from '@storybook/addon-links'
 import { withKnobs, text, boolean, number, object } from '@storybook/addon-knobs'
 
-import { CssCarousel } from '../lib/index'
+import { Carousel } from '../lib/index'
 import images from './data/slides'
 
 const CarouselStory = () => {
     console.log(images)
     return (
-        <>
-            <h2>Carousel</h2>
-            <CssCarousel width="600px">
-                {images.map((slide, index) => (
-                    <img key={index} src={slide.src} alt={slide.alt} width={slide.width} />
-                ))}
-            </CssCarousel>
-        </>
+        <Carousel width="100%">
+            {images.map((slide, index) => (
+                <img key={index} src={slide.src} alt={slide.alt} width={slide.width} />
+            ))}
+        </Carousel>
     )
 }
 
 storiesOf('Carousel', module).add('default', () => {
     return (
         <div>
+            <h2>Carousel</h2>
             <CarouselStory />
         </div>
     )
